backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.7
python-dotenv==1.1.0
google-adk==1.0.0
python-multipart==0.0.20